# Parquet reader performance