--- a/extension/parquet/column_reader.cpp
+++ b/extension/parquet/column_reader.cpp
@@ -514,15 +514,13 @@ bool ColumnReader::PrepareRead(idx_t read_now, data_ptr_t define_out, data_ptr_t
 
 	if (HasDefines()) {
 		D_ASSERT(defined_decoder);
-		if (defined_decoder->HasRepeatedBatch(read_now)) {
-			const auto repeated_value = defined_decoder->GetRepeatedBatch<uint8_t>(read_now);
-			if (repeated_value == MaxDefine()) {
-				return true; // Fast path: all values are valid
-			}
-			std::fill_n(define_out + result_offset, read_now, repeated_value);
-		} else {
-			defined_decoder->GetBatch<uint8_t>(repeat_out + result_offset, read_now);
+		const auto max_define = NumericCast<uint8_t>(MaxDefine());
+		if (!HasRepeats() && defined_decoder->HasRepeatedBatch<uint8_t>(read_now, max_define)) {
+			// Fast path: no repeats and all valid
+			defined_decoder->GetRepeatedBatch<uint8_t>(read_now, max_define);
+			return true;
 		}
+		defined_decoder->GetBatch<uint8_t>(define_out + result_offset, read_now);
 		return false;
 	}
 
@@ -624,9 +622,9 @@ void ColumnReader::DirectSelect(uint64_t num_values, data_ptr_t define_out, data
 
 	// we can only push the filter into the decoder if we are reading the ENTIRE vector in one go
 	if (read_now == to_read && encoding == ColumnEncoding::PLAIN) {
-		PrepareRead(read_now, define_out, repeat_out, 0);
-
-		PlainSelect(block, define_out, read_now, result, sel, approved_tuple_count);
+		const auto all_valid = PrepareRead(read_now, define_out, repeat_out, 0);
+		const auto define_ptr = all_valid ? nullptr : static_cast<uint8_t *>(define_out);
+		PlainSelect(block, define_ptr, read_now, result, sel, approved_tuple_count);
 
 		page_rows_available -= read_now;
 		FinishRead(num_values);
@@ -663,8 +661,8 @@ void ColumnReader::DirectFilter(uint64_t num_values, data_ptr_t define_out, data
 		} else {
 			// Push filter into dictionary directly
 			// read the defines/repeats
-			PrepareRead(read_now, define_out, repeat_out, 0);
-			auto define_ptr = HasDefines() ? static_cast<uint8_t *>(define_out) : nullptr;
+			const auto all_valid = PrepareRead(read_now, define_out, repeat_out, 0);
+			const auto define_ptr = all_valid ? nullptr : static_cast<uint8_t *>(define_out);
 			dictionary_decoder.Filter(define_ptr, read_now, result, filter, sel, approved_tuple_count);
 		}
 		page_rows_available -= read_now;
@@ -700,9 +698,9 @@ void ColumnReader::ApplyPendingSkips(data_ptr_t define_out, data_ptr_t repeat_ou
 
 	while (to_skip > 0) {
 		auto skip_now = ReadPageHeaders(to_skip);
-		PrepareRead(skip_now, define_out, repeat_out, 0);
+		const auto all_valid = PrepareRead(skip_now, define_out, repeat_out, 0);
 
-		auto define_ptr = HasDefines() ? static_cast<uint8_t *>(define_out) : nullptr;
+		const auto define_ptr = all_valid ? nullptr : static_cast<uint8_t *>(define_out);
 		switch (encoding) {
 		case ColumnEncoding::DICTIONARY:
 			dictionary_decoder.Skip(define_ptr, skip_now);
@@ -723,7 +721,7 @@ void ColumnReader::ApplyPendingSkips(data_ptr_t define_out, data_ptr_t repeat_ou
 			byte_stream_split_decoder.Skip(define_ptr, skip_now);
 			break;
 		default:
-			PlainSkip(*block, define_out, skip_now);
+			PlainSkip(*block, define_ptr, skip_now);
 			break;
 		}
 		page_rows_available -= skip_now;