--- a/extension/parquet/decoder/delta_length_byte_array_decoder.cpp
+++ b/extension/parquet/decoder/delta_length_byte_array_decoder.cpp
@@ -23,27 +23,56 @@ void DeltaLengthByteArrayDecoder::InitializePage() {
 
 void DeltaLengthByteArrayDecoder::Read(shared_ptr<ResizeableBuffer> &block_ref, uint8_t *defines, idx_t read_count,
                                        Vector &result, idx_t result_offset) {
+	if (defines) {
+		ReadInternal<true>(block_ref, defines, read_count, result, result_offset);
+	} else {
+		ReadInternal<false>(block_ref, defines, read_count, result, result_offset);
+	}
+}
+
+template <bool HAS_DEFINES>
+void DeltaLengthByteArrayDecoder::ReadInternal(shared_ptr<ResizeableBuffer> &block_ref, uint8_t *const defines,
+                                               const idx_t read_count, Vector &result, const idx_t result_offset) {
 	auto &block = *block_ref;
-	auto length_data = reinterpret_cast<uint32_t *>(length_buffer.ptr);
+	const auto length_data = reinterpret_cast<uint32_t *>(length_buffer.ptr);
 	auto result_data = FlatVector::GetData<string_t>(result);
 	auto &result_mask = FlatVector::Validity(result);
-	for (idx_t row_idx = 0; row_idx < read_count; row_idx++) {
-		auto result_idx = result_offset + row_idx;
-		if (defines && defines[result_idx] != reader.MaxDefine()) {
-			result_mask.SetInvalid(result_idx);
-			continue;
-		}
-		if (length_idx >= byte_array_count) {
+
+	if (!HAS_DEFINES) {
+		// Fast path: take this out of the loop below
+		if (length_idx + read_count >= byte_array_count) {
 			throw IOException(
 			    "DELTA_LENGTH_BYTE_ARRAY - length mismatch between values and byte array lengths (attempted "
 			    "read of %d from %d entries) - corrupt file?",
-			    length_idx, byte_array_count);
+			    length_idx + read_count, byte_array_count);
 		}
-		auto str_len = length_data[length_idx++];
-		block.available(str_len);
+		idx_t total_string_length = 0;
+		for (idx_t row_idx = 0; row_idx < read_count; row_idx++) {
+			total_string_length += length_data[length_idx + row_idx];
+		}
+		block.available(total_string_length);
+	}
+
+	for (idx_t row_idx = 0; row_idx < read_count; row_idx++) {
+		const auto result_idx = result_offset + row_idx;
+		if (HAS_DEFINES) {
+			if (defines[result_idx] != reader.MaxDefine()) {
+				result_mask.SetInvalid(result_idx);
+				continue;
+			}
+			if (length_idx >= byte_array_count) {
+				throw IOException(
+				    "DELTA_LENGTH_BYTE_ARRAY - length mismatch between values and byte array lengths (attempted "
+				    "read of %d from %d entries) - corrupt file?",
+				    length_idx, byte_array_count);
+			}
+			block.available(length_data[length_idx]);
+		}
+		const auto &str_len = length_data[length_idx++];
 		result_data[result_idx] = string_t(char_ptr_cast(block.ptr), str_len);
 		block.unsafe_inc(str_len);
 	}
+
 	StringColumnReader::ReferenceBlock(result, block_ref);
 }