--- a/extension/parquet/include/parquet_rle_bp_encoder.hpp
+++ b/extension/parquet/include/parquet_rle_bp_encoder.hpp
@@ -101,7 +101,9 @@ private:
 	idx_t rle_count;
 
 	//! BP stuff
-	static constexpr idx_t BP_BLOCK_SIZE = BitpackingPrimitives::BITPACKING_ALGORITHM_GROUP_SIZE;
+	static constexpr idx_t BP_BLOCK_SIZE = STANDARD_VECTOR_SIZE;
+	static_assert(BP_BLOCK_SIZE % BitpackingPrimitives::BITPACKING_ALGORITHM_GROUP_SIZE == 0,
+	              "BP_BLOCK_SIZE must be divisible by BITPACKING_ALGORITHM_GROUP_SIZE");
 	uint32_t bp_block[BP_BLOCK_SIZE] = {0};
 	uint32_t bp_block_packed[BP_BLOCK_SIZE] = {0};
 	idx_t bp_block_count;