Added a bunch of fast paths for when there are no `NULL` values, similar to what I did for the Parquet writer yesterday. Seems to improve performance by ~15% (when there are no `NULL` values). We perform this optimization once per scanned `Vector`, so even if there are some `NULL`s, the optimization can trigger.

Setup used to measure performance:
```sql
-- Load TPC-H SF 100
set preserve_insertion_order=false;
copy lineitem to 'lineitem_v1.parquet';
copy lineitem to 'lineitem_v2.parquet' (parquet_version v2);
select * from 'lineitem_v1.parquet' offset 600037901;
select * from 'lineitem_v2.parquet' offset 600037901;
```
Reading from `V1` has improved from ~5.5s to ~4.4s, and reading from `V2` has improved from ~4.2s to ~3.6s.